flue: address ask-bonk review findings (items 1-4)
1. Extract withConcurrency to lib/inproc-utils.ts; import from both
   code-review-inproc.ts and style-guide-inproc.ts instead of duplicating.

2. Fix read_repo_file tool UTF-8 decoding: use TextDecoder instead of bare
   atob() so non-ASCII content (em dashes, CJK, smart quotes) is not
   mojibake. Also encode path segments with encodeURIComponent, matching
   the getRepoFileContent fix already applied to github.ts.

3. Fix parseAddedLines: skip lines starting with backslash ('\ No newline
   at end of file') instead of treating them as context lines. Previously
   this incremented newLine, producing off-by-one line numbers for every
   subsequent added line in files without a trailing newline.

4. Fix stale comments: code-review-diff.ts, style-guide-inproc.ts, and
   wrangler.jsonc v4 comment still described the old in-orchestrator-DO
   architecture; updated to reflect the specialist-DO split.
